gdb_server: Add gdb_tdesc_path cmd qXfer read file
This patch add the command gdb_tdesc_path to set the path to gdb
XML target description file(s). Also the gdb qXfer:features:read
support is enabled if a path is set.
WARNING: the transfered filename from gdb is not parsed and it is
possible to read the first 4094 bytes from every accessible file
on the filesystem over this gdb connection. Feel free to fix this.

http://openocd.zylin.com/#/c/518/

Change-Id: I549c864ad66d185e96158faf22062de479556487

Signed-off-by: Mathias K <kesmtp@freenet.de>
Signed-off-by: Franck Jullien <franck.jullien@gmail.com>