target: add a list_type arg to target_get_gdb_reg_list
This new argument is used to control how much registers
must be given by the target. In case we have a lot of
registers and we don't want all of them in the g/G packets
we are now able to control how much registers to give
to target_get_gdb_reg_list.

Signed-off-by: Franck Jullien <franck.jullien@gmail.com>